% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fry_bubble.R
\name{bubble.comp.perc}
\alias{bubble.comp.perc}
\title{Bubble component in percentage}
\usage{
bubble.comp.perc(alpha.est, beta.est, v, num.obs)
}
\arguments{
\item{alpha.est}{is a parameter from hazard function}

\item{beta.est}{is a parameter from hazard function}

\item{v}{is equal to - ln[(1 - k)] > 0 and it is a key parameter of the Fry (2014) model}

\item{num.obs}{is the number of time observations}
}
\value{
bubble.component is the bubble component (in percentage)
}
\description{
This function computes the bubble component (in percentage) using eq.(16) on page 34 of Cheah and Fry (2015)
}
\details{
This function computes the bubble component (in percentage) using eq.(16) on page 34 of Cheah and Fry (2015)
}
\examples{
bubble.comp.perc(alpha.est=33.84404, beta.est=1.135961, v=0.546, num.obs=333)

}
\references{
Cheah, E.T., and Fry,J. (2015). Speculative bubbles in Bitcoin markets? An empirical investigation into the fundamental value of Bitcoin. Economics Letters, 130, 32-36.

Fry, J. (2014). Multivariate bubbles and antibubbles. The European Physical Journal B, 87(8), 174.
}
