% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PSY.R
\name{CV_BADF}
\alias{CV_BADF}
\title{Critical value sequences for the backward ADF (BADF) statistic.}
\usage{
CV_BADF(qe, m, T, r0, address, clust_number)
}
\arguments{
\item{qe}{is a q x 1 vector of quantiles}

\item{m}{is the number of replications for the simulated DGP}

\item{T}{is the number of simulated time steps}

\item{r0}{is the minimum window size fraction}

\item{address}{is a character object containing the address where the plot density of SADF test stat will be saved in png format}

\item{clust_number}{is the number of clusters for parallel computation}
}
\value{
quantile_badfs a q x 1 vector of simulated critical values for the sup ADF statistic
}
\description{
This function computes the critical values  for the backward ADF statistic  by Phillips et al. (2015)
}
\details{
This function computes the critical values for the backward ADF statistic by Phillips et al. (2015) and saves the plot density of the test statistic as a png file in a specified address
}
\examples{

\dontrun{
qe <-c(0.90,0.95,0.99)
address <- "D:/bubbletest/"
m <- 1000     # number of replications
system.time(aa<-CV_BADF(qe,m,100,0.190,address,clust_number=8))
}
}
