% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PSY.R
\name{CV_BSADF}
\alias{CV_BSADF}
\title{Critical values for the backward SADF (BSADF) statistic}
\usage{
CV_BSADF(qe, m, T, r0, address, clust_number, beta = 0.95)
}
\arguments{
\item{qe}{is a q x 1 vector of quantiles}

\item{m}{is the number of replications for the simulated DGP}

\item{T}{is the number of simulated time steps}

\item{r0}{is the minimum window size fraction}

\item{address}{is a character object containing the address where the plot density of SADF test stat will be saved in png format}

\item{clust_number}{is the number of clusters for parallel computation}

\item{beta}{is the significance level for the critical values for the BSADF sequence}
}
\value{
results is a list containing quantile_bsadf (a q x 1 vector of simulated critical values for the BSADF statistic) and bsadf_cv (a (T-swindow0+1) vector containing the critical values for the BSADF sequence)
}
\description{
This function computes the critical values for the backward SADF (BSADF) statistic  by Phillips et al. (2015)
}
\details{
This function computes the simulated critical values for the backward SADF (BSADF) statistic by Phillips et al. (2015) and saves the plot density of the test statistic as a png file in a specified address.
The main difference between the SADF and GSADF tests on one side and BADF BSADF tests on the other side, lies in the data generating process:
the SADF and GSADF tests use a random walk with an overall drift value of 1/T, while the
the BADF BSADF tests use a random walk with an drift value of 1/r2 for each r2
}
