% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PSY.R
\name{DGP}
\alias{DGP}
\title{Data generating process for simulations: random walk with drift 1/n}
\usage{
DGP(n, niter)
}
\arguments{
\item{n}{is the number of simulated time steps}

\item{niter}{is the number of replications for the simulated DGP}
}
\value{
y a n x niter matrix of simulated data, where each colum represents a single simulated process
}
\description{
This function simulates a random walk with drift 1/n, which is used when computing
the tests for bubbles by Phillips et al. (2015)
}
\details{
This function simulates a n x 1 vector following a a random walk with drift 1/n,
and this operation is repeated niter times
}
