% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watanabe_functions.R
\name{est.war1.from.arp}
\alias{est.war1.from.arp}
\title{Simulate data from a AR(p) model and estimate the particular AR(1) model by Watanabe et al.(2007a,b) using a simplified method}
\usage{
est.war1.from.arp(ar.par, step.win)
}
\arguments{
\item{ar.par}{is a a p x 1 vector containing the AR parameters for the simulation of the AR(p)}

\item{step.win}{is the number of simulated time steps}
}
\value{
vec.1r.1 is a 1 x 1 numeric scalar, which is the estimated omega(i,ti) parameter
}
\description{
This function simulates data from a AR(p) model and estimate the particular AR(1) model by Watanabe et al.(2007a,b) using a simplified method
}
\details{
This function simulates data from a AR(p) model and estimate the particular AR(1) model by Watanabe et al.(2007a,b) using a simplified
method which does not required nonlinear optimization,
 and returns the estimated omega(i,ti) parameter
}
