% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watanabe_functions.R
\name{est.war1}
\alias{est.war1}
\title{Estimate the particular AR(1) model by Watanabe et al.(2007a,b) using a simplified method}
\usage{
est.war1(x)
}
\arguments{
\item{x}{is a T x 1 numeric data vector}
}
\value{
vec.1r.1 is a 1 x 1 numeric scalar, which is the estimated omega(i,ti) parameter
}
\description{
This function estimates the particular AR(1) model by Watanabe et al.(2007a,b) using a simplified method
}
\details{
This function estimates the particular AR(1) model by Watanabe et al.(2007a,b) using a simplified method
which does not required nonlinear optimization and returns the estimated omega(i,ti) parameter
}
