% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PSY.R
\name{GSADF_Y}
\alias{GSADF_Y}
\title{GSADF test statistic and the BSADF sequence for a real time series data}
\usage{
GSADF_Y(y, r0, lag, select, clust_number)
}
\arguments{
\item{y}{is a T x 1 numeric vector}

\item{r0}{is the minimum window size fraction}

\item{lag}{is the number of lags in the ADF regression}

\item{select}{is a character object choosing either a "Fixed" number of lags in the ADF regression or by using the the "BIC"/"AIC" criteria}

\item{clust_number}{is the number of clusters for parallel computation}
}
\value{
results is a list containing the GSADF test statistic and the BSADF sequence
}
\description{
This function computes the GSADF test statistic and the BSADF sequence for a real time series data
}
\details{
This function computes the GSADF test statistic and BSADF sequence for a real time series data. The minimum window size fraction suggested by PSY(2015) is r0=0.01+1.8/sqrt(length(y))
}
\examples{

\dontrun{
#Replicate the empirical analysis in Table 8 by PSY(2015)

# Read file
path.bit <- system.file("extdata", "SP_DV.csv", package = "bubble")
SP_DV <- read.table(path.bit,  head = TRUE, sep = ";", fill=T, stringsAsFactors = FALSE)
SP_DV <- na.omit( SP_DV[,"PV.ratio", drop=F])
r0=0.01+1.8/sqrt(length(SP_DV[,1]))
gsadf_test <-GSADF_Y(SP_DV[,1], r0, 6 , "BIC", 8)

qe <-c(0.90,0.95,0.99)    #quantiles
m <- 1000
T=length(SP_DV[,1])
clust_number=7
cv_gsadf<- CV_GSADF(qe,m,T,r0, clust_number=7, address)
cat(gsadf_test$gsadf, cv_gsadf, sep = "\\t")
}
}
