% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPPL_basic.R
\name{lppl.resids.from.basic.estimate}
\alias{lppl.resids.from.basic.estimate}
\title{Compute the residuals given the parameters estimated with the function lppl_estimate}
\usage{
lppl.resids.from.basic.estimate(x, par_est)
}
\arguments{
\item{x}{is a T x 1 numeric data vector}

\item{par_est}{is a 7 x 1 vector of parameters estimated with the function lppl_estimate
in this order: ["bet" "ome" "phi" "tc"  "A"   "B"   "C"]}
}
\value{
et is a vector of residuals
}
\description{
This function computes the residuals given the parameters estimated with the function lppl_estimate
}
\details{
This function computes the residuals given the parameters estimated with the function lppl_estimate
}
\examples{

 \dontrun{
 tparm <- c(0.353689, 9.154368, 2.074608, 7.166421,-0.434324, 0.035405, 0.000071, 30)
 aa <- lppl_simulate(500,tparm)
 bb <- lppl_estimate(aa)
 bb
 resids <- lppl.resids.from.basic.estimate(aa,bb)
 }

}
