% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPPL_basic.R
\name{lppl_estimate}
\alias{lppl_estimate}
\title{Estimation of the LPPL model using a nonlinear optimization}
\usage{
lppl_estimate(x)
}
\arguments{
\item{x}{is a T x 1 numeric data vector}
}
\value{
par_est is a 7 x 1 vector of estimated parameters
}
\description{
This function estimates a LPPL model using a nonlinear optimization
}
\details{
This function estimates the LPPL model by Johansen, Ledoit, and Sornette (2000) using the original (two-step) nonlinear optimization, see section 4.1 in Geraskin and Fantazzini (2013) for a compact review.
The returned parameter vector contained the following parameters:

par_est[1] = beta

par_est[2] = omega

par_est[3] = phi

par_est[4] = tc (i.e. the critical time)

par_est[5] = A

par_est[6] = B

par_est[7] = C

We remark that this estimation method is not recommended, due to the frequent presence of many
local minima of the cost function where the minimization algorithm can get trapped.
It was included in this package for historical (and teaching) reasons.
}
\examples{

 \dontrun{
 tparm=c(0.353689, 9.154368, 2.074608, 7.166421,-0.434324, 0.035405, 0.000071, 530)
 aa=lppl_simulate(500,tparm)

 bb=lppl_estimate(aa); bb;
 }

}
