% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPPL_3step.R
\name{lppl_estimate_rob_2s}
\alias{lppl_estimate_rob_2s}
\title{2nd step LPPL estimation procedure  by Geraskin and Fantazzini (2013) and Fantazzini(2016)}
\usage{
lppl_estimate_rob_2s(x, par, max.win.tc = 0.1)
}
\arguments{
\item{x}{is a T x 1 data vector}

\item{par}{is a 6 x 1 vector containing the parameters estimated in the 1st step [beta, omega, A, B, C1, C2] and which are kept fixed in the 2nd step}

\item{max.win.tc}{is a scalar (in percentage terms) used to set the starting value for the critical time tc}
}
\value{
par_est2 is a 1 x 1 scalar containing the estimated parameter tc
}
\description{
This function performs the 2nd step of the LPPL estimation procedure
by Geraskin and Fantazzini (2013) and Fantazzini (2016)
}
\details{
This function performs the 2nd step of the LPPL estimation procedure
by Geraskin and Fantazzini (2013) and Fantazzini (2016) using the LPPL formula
 by Filimonov and Sornette (2013):
Keeping fixed the LPPL parameters  [beta, omega, A, B, C1, C2] computed in the
first stage, the critical time tc is estimated in a second step by using a
 quasi-Newton method algorithm.
}
