% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPPL_3step.R
\name{lppl_estimate_rob_3all}
\alias{lppl_estimate_rob_3all}
\title{3-step LPPL estimation procedure by Geraskin and Fantazzini (2013) and Fantazzini (2016)}
\usage{
lppl_estimate_rob_3all(x, par_start = c(bet = 0.5, ome = 6, A =
  log(x[NROW(x)]), B = 0, C1 = 0, C2 = 0), max.win.tc = 0.1)
}
\arguments{
\item{x}{is a T x 1 data vector}

\item{par_start}{is a 6 x 1 vector of starting values for the parameters to be estimated in the first step}

\item{max.win.tc}{is a scalar setting the max window size (in percentage terms) used to fix the critical time tc in the 1st step, and to set the starting value for tc in the 2nd step}
}
\value{
results is a list containing the following objects:

- a vector named param.JLS containing the estimated parameters according to the original LPPL structure by Johansen et al. (2001): [beta, omega, phi, A, B, C, tc]

- a vector named param.FS containing the estimated parameters according to the LPPL structure by Filimonov and Sornette (2013): [beta, omega, A, B, C1, C2, tc]

- the (scalar) KPSS test statistic computed with the LPPL residuals to check their stationarity

- the crash hazard rate computed according to Bothmer and Meister (2003)

- the Relative Error of the model fit: mean( (Y-Yfit)/Yfit )

- the model residuals
}
\description{
This function performs all the 3 steps of the LPPL estimation procedure  by Geraskin and Fantazzini (2013) and Fantazzini (2016)
}
\details{
This function performs all the 3 steps of the LPPL estimation procedure  by Geraskin and Fantazzini (2013) and Fantazzini (2016) using the LPPL
formula by Filimonov and Sornette (2013). See the functions for each of these estimation steps for more details.
}
\examples{

\dontrun{
 tparm=c(0.35, 4.15, 2.07, 7.16,-0.43, 0.035, 0.00007, 530)
 aa=lppl_simulate(500,tparm)
 est.all<-lppl_estimate_rob_3all(aa)
 est.all
}
}
