% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPPL_basic.R
\name{lppl_simulate}
\alias{lppl_simulate}
\title{Simulate a LPPL model}
\usage{
lppl_simulate(T = 500, true_parm)
}
\arguments{
\item{T}{is the number of simulated time steps}

\item{true_parm}{is a a 8 x 1 vector containing the parameters for the simulation}
}
\value{
y a T  x 1 vector of simulated data
}
\description{
This function simulates a LPPL model
}
\details{
This function simulates a LPPL model with parameter vector given by true_param and make a plot of it. More specifically, the LPPL parameters in the true_param vector are given below:

true_parm[1] = beta

true_parm[2] = omega

true_parm[3] = phi

true_parm[4] = A

true_parm[5] = B

true_parm[6] = C

true_parm[7] = sigma (i.e. the error term variance)

true_parm[8] = the critical time tc
}
\examples{

 \dontrun{
 tparm=c(0.353689, 9.154368, 2.074608, 7.166421,-0.434324, 0.035405, 0.000071, 530)
 aa=lppl_simulate(500,tparm)
 }

}
