% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPPL_basic.R
\name{lppl_simulate_boot}
\alias{lppl_simulate_boot}
\title{Simulate a LPPL model using bootstrap resampling of residuals}
\usage{
lppl_simulate_boot(T = 500, true_parm, resids)
}
\arguments{
\item{T}{is the number of simulated time steps}

\item{true_parm}{is a a 8 x 1 vector containing the parameters for the simulation}

\item{resids}{is a T x 1 vector of residuals from a previous LPPL estimation}
}
\value{
y a T x 1 vector of simulated data
}
\description{
This function simulates a LPPL model using bootstrap resampling of residuals
}
\details{
This function simulates a LPPL model with parameter vector given by true_param,
and error terms given by the bootstrp resampling of residuals from a previous LPPL estimation.
The simulated data are then plotted. More specifically,
the LPPL parameters in the true_param vector are given below:

true_parm[1] = beta

true_parm[2] = omega

true_parm[3] = phi

true_parm[4] = A

true_parm[5] = B

true_parm[6] = C

true_parm[7] = the critical time tc
}
\examples{

 \dontrun{
 tparm=c(0.353689, 9.154368, 2.074608, 7.166421,-0.434324, 0.035405, 0.000071, 530)
 aa=lppl_simulate(500,tparm)
 bb=lppl_estimate(aa); bb;
 resids <- lppl.resids.from.basic.estimate(aa,bb)
 tparm.boot=c(0.353689, 9.154368, 2.074608, 7.166421,-0.434324, 0.035405, 530)
 aa=lppl_simulate_boot(500,tparm.boot, resids)
 }

}
