% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPPL_3step.R
\name{lppl_simulate_FS}
\alias{lppl_simulate_FS}
\title{Simulate a LPPL model using the formula in Filimonov and Sornette (2013)}
\usage{
lppl_simulate_FS(T = 500, true_parm)
}
\arguments{
\item{T}{is the number of simulated time steps}

\item{true_parm}{is a a 8 x 1 vector containing the parameters for the simulation}
}
\value{
y a T  x 1 vector of simulated data
}
\description{
This function simulates a LPPL model using the formula in Filimonov and Sornette (2013)
}
\details{
This function simulates a LPPL model using the formula in Filimonov and Sornette (2013),
 with parameter vector given by true_param and make a plot of it. More specifically,
 the LPPL parameters in the true_param vector are given below:

true_parm[1] = beta

true_parm[2] = omega

true_parm[3] = A

true_parm[4] = B

true_parm[5] = C1

true_parm[6] = C2

true_parm[7] = the critical time tc

true_parm[8] = sigma (i.e. the error term variance)
}
\examples{

 \dontrun{
 tparm=c(0.35, 4.15, 7.16,-0.43, -0.016, 0.031, 530, 0.00007)
 a1=lppl_simulate_FS(500,tparm)
 est.all<-lppl_estimate_rob_3all(a1)
 est.all$param.FS
 }

}
