% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPPL_3step.R
\name{lppl_simulate_FS_boot}
\alias{lppl_simulate_FS_boot}
\title{Simulate a LPPL model using bootstrap resampling of residuals and the formula in Filimonov and Sornette (2013)}
\usage{
lppl_simulate_FS_boot(T = 500, true_parm, resids)
}
\arguments{
\item{T}{is the number of simulated time steps}

\item{true_parm}{is a a 7 x 1 vector containing the parameters for the simulation}

\item{resids}{is a T x 1 vector of residuals from a previous LPPL estimation}
}
\value{
y a T x 1 vector of simulated data
}
\description{
This function simulates a LPPL model using bootstrap resampling of residuals and the formula in Filimonov and Sornette (2013)
}
\details{
This function simulates a LPPL model using the formula in Filimonov and Sornette (2013),
error terms given by the bootstrp resampling of residuals from a previous LPPL estimation,
 and with a parameter vector given by true_param. The simulated data are then plotted. More specifically,
 the LPPL parameters in the true_param vector are given below:

true_parm[1] = beta

true_parm[2] = omega

true_parm[3] = A

true_parm[4] = B

true_parm[5] = C1

true_parm[6] = C2

true_parm[7] = the critical time tc
}
\examples{

 \dontrun{
 tparm=c(0.35, 4.15, 7.16,-0.43, -0.016, 0.031, 530, 0.00007)
 a1=lppl_simulate_FS(500,tparm)
 est.all<-lppl_estimate_rob_3all(a1)
 resids=est.all$resids
 tparm.boot=c(0.35, 4.15, 7.16,-0.43, -0.016, 0.031, 530)
 aa=lppl_simulate_FS_boot(500,tparm.boot, resids)
 }

}
