% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\docType{data}
\name{NodeSession}
\alias{NodeSession}
\title{NodeJS Session}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NodeSession
}
\arguments{
\item{bin}{Path to NodeJs bin directory, if \code{NULL} then bubble 
attemtpts to find the directory with \code{\link{find_node}}.}

\item{params}{Additional parameters to pass to the initialisation.}

\item{code}{The code to evaluate.}

\item{wait}{Whether to re-attempt to evaluate if it first fails.}

\item{print}{Whether to print the result to the R console.}

\item{...}{Bare name of objects to retreive.}
}
\description{
Launch a NodeJS Session
}
\details{
Initialise a NodeJs session

Terminate a NodeJs session

Evaluate NodeJs code

Retrieve NodeJs objects

Retrieve NodeJs state

Kill NodeJs

Terminate NodeJs
}
\section{Fields}{

\describe{
\item{\code{bin}}{Path to NodeJs bin directory.}

\item{\code{handle}}{A handle as returned by \link[subprocess]{spawn_process}.}
}}

\examples{
\dontrun{
n <- NodeSession$new()
n$eval("17 + 29")
}
\dontrun{
n <- NodeSession$new()
n$eval("17 + 29")
}
\dontrun{
n <- NodeSession$new()
n$eval("var x = 12")
n$eval("var y = 17")
n$get(x, y)
}
\dontrun{
n <- NodeSession$new()
n$state()
n$kill()
n$state()
}
\dontrun{
n <- NodeSession$new()
n$kill()
n$state()
}
\dontrun{
n <- NodeSession$new()
n$terminate()
n$state()
}
}
\keyword{datasets}
