% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{Npm}
\alias{Npm}
\title{Node Package Manager}
\description{
Create and interact with npm.
}
\examples{

## ------------------------------------------------
## Method `Npm$new`
## ------------------------------------------------

\dontrun{Npm$new()}

## ------------------------------------------------
## Method `Npm$cmd`
## ------------------------------------------------

\dontrun{Npm$new()$cmd("ls")}

## ------------------------------------------------
## Method `Npm$install`
## ------------------------------------------------

\dontrun{Npm$new()$install("browserify")}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bin}}{Path to npm bin directory.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Npm$new()}}
\item \href{#method-cmd}{\code{Npm$cmd()}}
\item \href{#method-install}{\code{Npm$install()}}
\item \href{#method-clone}{\code{Npm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Npm$new(bin = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bin}}{Path to npm bin directory.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise npm
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Npm$new()}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cmd"></a>}}
\subsection{Method \code{cmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Npm$cmd(command, args = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{command}}{Command to execute.}

\item{\code{args}}{Additional arguments and flags.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Execute command
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Npm$new()$cmd("ls")}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-install"></a>}}
\subsection{Method \code{install()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Npm$install(package = NULL, global = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{package}}{Name of npm package to install. If \code{NULL}
install dependencies from lock file.}

\item{\code{global}}{Whether to install globally.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Install dependencies
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Npm$new()$install("browserify")}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Npm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
