% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{NpmSession}
\alias{NpmSession}
\title{Node Package Manager}
\description{
Create and interact with npm.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NpmSession$new()}}
\item \href{#method-install}{\code{NpmSession$install()}}
\item \href{#method-clone}{\code{NpmSession$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NpmSession$new(bin = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-install"></a>}}
\subsection{Method \code{install()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NpmSession$install(package = NULL, global = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NpmSession$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
