% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PSY.R
\name{size_GSADF}
\alias{size_GSADF}
\title{Size of the generalized sup ADF (GSADF) statistic.}
\usage{
size_GSADF(qe, m, T, r0, lag, select, cv, clust_number)
}
\arguments{
\item{qe}{is a q x 1 vector of quantiles}

\item{m}{is the number of replications for the simulated DGP}

\item{T}{is the number of simulated time steps}

\item{r0}{is the minimum window size fraction}

\item{lag}{is the number of lags in the ADF regression}

\item{select}{is a character object choosing either a "fixed" number of lags in the ADF regression or by using the the "BIC"/"AIC" criteria}

\item{cv}{is the critical value of the GSADF test}

\item{clust_number}{is the number of clusters for parallel computation}
}
\value{
size is the size of the test
}
\description{
This function computes the size of the generalized sup ADF (GSADF) statistic to replicate the results in Table 2 by Phillips et al. (2015)
}
\details{
This function computes the size of the generalized sup ADF (GSADF) statistic to replicate the results in Table 2 by Phillips et al. (2015),
}
\examples{

\dontrun{
# Replicate sizes of the GSADF test in Table 2 by PSY(2015)
qe <-c(0.90,0.95,0.99)
m <- 1000           # number of replications
clust_number <- 7

table2_GSADF <- matrix(0, 5, 3)
table2_GSADF[1,1] <- size_GSADF(qe, m, 100, 0.190, 0 , "Fixed", 2.00, clust_number)
table2_GSADF[2,1] <- size_GSADF(qe, m, 200, 0.137, 0 , "Fixed", 2.08, clust_number)
table2_GSADF[3,1] <- size_GSADF(qe, m, 400, 0.100, 0 , "Fixed", 2.20, clust_number)
table2_GSADF[4,1] <- size_GSADF(qe, m, 800, 0.074, 0 , "Fixed", 2.34, clust_number)
table2_GSADF[5,1] <- size_GSADF(qe, m, 1600,0.055, 0 , "Fixed", 2.41, clust_number)

table2_GSADF[1,2] <- size_GSADF(qe, m, 100, 0.190, 3 , "Fixed", 2.00, clust_number)
table2_GSADF[2,2] <- size_GSADF(qe, m, 200, 0.137, 3 , "Fixed", 2.08, clust_number)
table2_GSADF[3,2] <- size_GSADF(qe, m, 400, 0.100, 3 , "Fixed", 2.20, clust_number)
table2_GSADF[4,2] <- size_GSADF(qe, m, 800, 0.074, 3 , "Fixed", 2.34, clust_number)
table2_GSADF[5,2] <- size_GSADF(qe, m, 1600,0.055, 3 , "Fixed", 2.41, clust_number)

table2_GSADF[1,3] <- size_GSADF(qe, m, 100, 0.190, 6 , "BIC", 2.00, clust_number)
table2_GSADF[2,3] <- size_GSADF(qe, m, 200, 0.137, 6 , "BIC", 2.08, clust_number)
table2_GSADF[3,3] <- size_GSADF(qe, m, 400, 0.100, 6 , "BIC", 2.20, clust_number)
table2_GSADF[4,3] <- size_GSADF(qe, m, 800, 0.074, 6 , "BIC", 2.34, clust_number)
table2_GSADF[5,3] <- size_GSADF(qe, m, 1600,0.055, 6 , "BIC", 2.41, clust_number)
}
}
