% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PSY.R
\name{size_SADF}
\alias{size_SADF}
\title{Size of the  sup ADF test}
\usage{
size_SADF(qe, m, T, r0, lag, select, cv, clust_number)
}
\arguments{
\item{qe}{is a q x 1 vector of quantiles}

\item{m}{is the number of replications for the simulated DGP}

\item{T}{is the number of simulated time steps}

\item{r0}{is the minimum window size fraction}

\item{lag}{is the number of lags in the ADF regression}

\item{select}{is a character object choosing either a "fixed" number of lags in the ADF regression or by using the the "BIC"/"AIC" criteria}

\item{cv}{is the critical value of the SADF test}

\item{clust_number}{is the number of clusters for parallel computation}
}
\value{
size is the size of the test
}
\description{
This function computes the size of the sup ADF statistic to replicate the results in Table 2 by Phillips et al. (2015)
}
\details{
This function computes the size of the sup ADF statistic to replicate the results in Table 2 by Phillips et al. (2015),
}
\examples{

 \dontrun{
 # Replicate sizes of the SADF test in Table 2 by PSY(2015)
 qe <-c(0.90,0.95,0.99)
 m <- 1000           # number of replications

 table2_SADF <- matrix(0, 5, 3)
 table2_SADF[1,1] <- size_SADF(qe, m, 100, 0.190, 0 , "Fixed", 1.30, clust_number=8)
 table2_SADF[2,1] <- size_SADF(qe, m, 200, 0.137, 0 , "Fixed", 1.40, clust_number=8)
 table2_SADF[3,1] <- size_SADF(qe, m, 400, 0.100, 0 , "Fixed", 1.49, clust_number=8)
 table2_SADF[4,1] <- size_SADF(qe, m, 800, 0.074, 0 , "Fixed", 1.53, clust_number=8)
 table2_SADF[5,1] <- size_SADF(qe, m, 1600,0.055, 0 , "Fixed", 1.57, clust_number=8)

 table2_SADF[1,2] <- size_SADF(qe, m, 100, 0.190, 3 , "Fixed", 1.30, clust_number=8)
 table2_SADF[2,2] <- size_SADF(qe, m, 200, 0.137, 3 , "Fixed", 1.40, clust_number=8)
 table2_SADF[3,2] <- size_SADF(qe, m, 400, 0.100, 3 , "Fixed", 1.49, clust_number=8)
 table2_SADF[4,2] <- size_SADF(qe, m, 800, 0.074, 3 , "Fixed", 1.53, clust_number=8)
 table2_SADF[5,2] <- size_SADF(qe, m, 1600,0.055, 3 , "Fixed", 1.57, clust_number=8)

 table2_SADF[1,3] <- size_SADF(qe, m, 100, 0.190, 6 , "BIC", 1.30, clust_number=8)
 table2_SADF[2,3] <- size_SADF(qe, m, 200, 0.137, 6 , "BIC", 1.40, clust_number=8)
 table2_SADF[3,3] <- size_SADF(qe, m, 400, 0.100, 6 , "BIC", 1.49, clust_number=8)
 table2_SADF[4,3] <- size_SADF(qe, m, 800, 0.074, 6 , "BIC", 1.53, clust_number=8)
 table2_SADF[5,3] <- size_SADF(qe, m, 1600,0.055, 6 , "BIC", 1.57, clust_number=8)
 }
}
