% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{json_conversion}
\alias{json_conversion}
\alias{as_json_string}
\alias{as_json_file}
\title{Convert to json}
\usage{
as_json_string(name, value, type)

as_json_file(name, value, type)
}
\arguments{
\item{name}{Name of variable.}

\item{value}{Value to convert.}

\item{type}{Variable type.}
}
\description{
Convert to json
}
\keyword{internal}
