% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{json_conversion}
\alias{json_conversion}
\alias{as_json}
\alias{json_var}
\title{Convert to json}
\usage{
as_json(value)

json_var(name, value)
}
\arguments{
\item{value}{Value to convert.}

\item{name}{Name of variable.}
}
\description{
Convert to json
}
\keyword{internal}
