% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\docType{data}
\name{NodeSession}
\alias{NodeSession}
\title{NodeJS Session}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NodeSession
}
\description{
Launch a NodeJS Session
}
\keyword{datasets}
