% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{NodeSession}
\alias{NodeSession}
\title{NodeJS Session}
\description{
Launch a NodeJS Session
}
\examples{

## ------------------------------------------------
## Method `NodeSession$new`
## ------------------------------------------------

\dontrun{
n <- NodeSession$new()
n$eval("17 + 29")
}

## ------------------------------------------------
## Method `NodeSession$eval`
## ------------------------------------------------

\dontrun{
n <- NodeSession$new()
n$eval("17 + 29")
}

## ------------------------------------------------
## Method `NodeSession$assign`
## ------------------------------------------------

\dontrun{
n <- NodeSession$new()
n$assign("cars", cars)
n$get(cars)
}

## ------------------------------------------------
## Method `NodeSession$get`
## ------------------------------------------------

\dontrun{
n <- NodeSession$new()
n$eval("var x = 12")
n$eval("var y = 17")
n$get(x, y)
}

## ------------------------------------------------
## Method `NodeSession$state`
## ------------------------------------------------

\dontrun{
n <- NodeSession$new()
n$state()
n$kill()
n$state()
}

## ------------------------------------------------
## Method `NodeSession$kill`
## ------------------------------------------------

\dontrun{
n <- NodeSession$new()
n$kill()
n$state()
}

## ------------------------------------------------
## Method `NodeSession$terminate`
## ------------------------------------------------

\dontrun{
n <- NodeSession$new()
n$terminate()
n$state()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bin}}{Path to NodeJs bin directory.}

\item{\code{handle}}{A handle as returned by \link[subprocess]{spawn_process}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NodeSession$new()}}
\item \href{#method-finalize}{\code{NodeSession$finalize()}}
\item \href{#method-eval}{\code{NodeSession$eval()}}
\item \href{#method-assign}{\code{NodeSession$assign()}}
\item \href{#method-get}{\code{NodeSession$get()}}
\item \href{#method-state}{\code{NodeSession$state()}}
\item \href{#method-kill}{\code{NodeSession$kill()}}
\item \href{#method-terminate}{\code{NodeSession$terminate()}}
\item \href{#method-clone}{\code{NodeSession$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$new(bin = NULL, params = "-i")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bin}}{Path to NodeJs bin directory, if \code{NULL} then bubble 
attemtpts to find the directory with \code{\link{find_node}}.}

\item{\code{params}}{Additional parameters to pass to the initialisation.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise a NodeJs session
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
n <- NodeSession$new()
n$eval("17 + 29")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$finalize()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Terminate a NodeJs session
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval"></a>}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$eval(code, wait = TRUE, print = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{The code to evaluate.}

\item{\code{wait}}{Whether to re-attempt to evaluate if it first fails.}

\item{\code{print}}{Whether to print the result to the R console.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Evaluate NodeJs code
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
n <- NodeSession$new()
n$eval("17 + 29")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assign"></a>}}
\subsection{Method \code{assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$assign(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of variable to create.post}

\item{\code{value}}{Value to assign to variable.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create NodeJs objects
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
n <- NodeSession$new()
n$assign("cars", cars)
n$get(cars)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$get(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Bare name of objects to retreive.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Retrieve NodeJs objects
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
n <- NodeSession$new()
n$eval("var x = 12")
n$eval("var y = 17")
n$get(x, y)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-state"></a>}}
\subsection{Method \code{state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$state()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Retrieve NodeJs state
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
n <- NodeSession$new()
n$state()
n$kill()
n$state()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kill"></a>}}
\subsection{Method \code{kill()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$kill()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Kill NodeJs
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
n <- NodeSession$new()
n$kill()
n$state()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-terminate"></a>}}
\subsection{Method \code{terminate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$terminate()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Terminate NodeJs
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
n <- NodeSession$new()
n$terminate()
n$state()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
