% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{NodeSession}
\alias{NodeSession}
\title{NodeJS Session}
\description{
Launch a NodeJS Session
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NodeSession$new()}}
\item \href{#method-finalize}{\code{NodeSession$finalize()}}
\item \href{#method-eval}{\code{NodeSession$eval()}}
\item \href{#method-get}{\code{NodeSession$get()}}
\item \href{#method-state}{\code{NodeSession$state()}}
\item \href{#method-kill}{\code{NodeSession$kill()}}
\item \href{#method-terminate}{\code{NodeSession$terminate()}}
\item \href{#method-clone}{\code{NodeSession$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$new(bin = NULL, params = "-i")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval"></a>}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$eval(code, wait = TRUE, print = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$get(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-state"></a>}}
\subsection{Method \code{state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kill"></a>}}
\subsection{Method \code{kill()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$kill()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-terminate"></a>}}
\subsection{Method \code{terminate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$terminate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeSession$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
