% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\docType{data}
\name{Npm}
\alias{Npm}
\title{Node Package Manager}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Npm
}
\arguments{
\item{bin}{Path to npm bin directory.}

\item{command}{Command to execute.}

\item{...}{Additional arguments and flags.}

\item{package}{Name of npm package to install. If \code{NULL}
install dependencies from lock file.}

\item{global}{Whether to install globally.}

\item{package}{Name of npm package to install. If \code{NULL}
install dependencies from lock file.}

\item{global}{Whether to install globally.}
}
\description{
Create and interact with npm.
}
\details{
Initialise npm

Execute command

Install packages

Uninstall packages
}
\section{Fields}{

\describe{
\item{\code{bin}}{Path to npm bin directory.}
}}

\examples{
\dontrun{Npm$new()}
\dontrun{Npm$new()$cmd("ls")}
\dontrun{Npm$new()$install("browserify")}
\dontrun{Npm$new()$install("browserify")}
}
\keyword{datasets}
